♻️. Create A Box Type Matching Asset & Detail
Before there were two arrays on `Account` for assets(`Asset`) and asset
details(`AssetInformation`), and if it was needed to access one using the
other, the corresponding array were filtered, which may not be performance-
wise since it can contains lots of items.
Now, the array holding asset details(`AssetInformation`) is converted to
one of a third type(`CompoundAsset`) containing the asset and detail. Also,
there is a exclusive dictionary to map an asset id with the index of the
compound asset. It will improve the performance if it is needed to find
the compound asset with an id.