🐛 Set The Correct Order For Newly-Added Accounts
Previously, the preferred order of the last account added to the app was
being used, but the user could update the order before adding a new one,
then the wrong order would be set to the newly-added account because of
the wrong selection of the last account. Now, the local accounts are
sorted by their preferred order, then the last one in the list is being
used to determine the next order for the newly-added account.