🐛 Fix account ordering interaction issue
Issue: If you don't move the cell to other index path, the delegate method won't be triggered, which causes `movingAccountOrderingPreviewCellIndexPath` to stay as nil. When it stays nil we guard with nil check, therefore we can't customize the cell back to non moving state.

Fixed it by setting the `movingAccountOrderingPreviewCellIndexPath` in `began` state