🐛 Dismiss the last toast on app resigning active
The problem was that when any toast message were presented, if the app
entered background and came back, the new toast messages were queued up,
weren't presenting again till the app is relaunched. It is mostly fixed
on 'Macaroon' library. In the app, we are just dismissing the presented
toast message on app entering background.