🐛 Fix the formatting in Europe region locales
The currency symbol is on the right side of the formatting numbers if the
user selects a locale from the Europe region. However, our previous
approach was just overriding the currency symbol when we had added
abbreviation to the number because we set the `positiveSuffix` and
`negativeSuffix` properties of the formatter instance, and the system
was also set them internally for the currency symbol. Once they are set,
the system stops to update them even if the currency code or symbol are
set in the formatter instance. Now, the abbreviation is being inserted
to the right position manually.