💄 Fix UI glitches on the system auto-bounces
In 'Home' and 'Account Assets List' screens, the system triggers
auto-bouncing back when applying a snapshot if the user pulls down
much more than the content starting position, and since there is a
custom-colored layer at the top, it caused UI glitches revealing it.
It is fixed with a workaround by increasing the minimum height of the
layer, so it isn't visible there.