🐛 Fix WalletConnect session disconnect issue
Issue: When user deletes all data from the app, we reset all wallet connect sessions, but we were not disconnecting from them. So dApp is able communicate with the app when we add the same account back.

Fix: It is fixed by iterating through all wallet connect sessions and disconnecting them when delete all action is performed.
Co-Authored-By: sakkaras <sakkaras@users.noreply.github.com>