🐛 Fix copy/paste feature for mnemonics
When the user tried to copy/paste a set of mnemonics, the text had to be
a certain format to be able to handle it properly, i.e. it had to have
one space between words, or each word had to have a comma, and the space
next to it. Now, a new parser is implemented to be able to support
whitespace, new line, any punctuations, and any symbols between the words.