✨ Implement configuration of PrimaryListItemView for asset list item
Implemented view models and themes related to Algo and Asset list items.
Used two separate view model for Algo and Asset to avoid complexity.
Confirmed Hashable protocol so that it can be used on the lists with Diffable Data Sources when it's needed.
Created AssetListItemCell to display it on the needed places.