✨ Change AccountAssetList to work with newly added Asset List Items
- Account asset list was depending on AssetPreviewModel and Asset Adapter.
Moved this to make it work with Asset List Item View Models for Algo and other
assets. Changed the related data source, layout and controllers accordingly.
- Asset sorting mechanisms was working on AssetPreviewModel. Changed the all
types of sorting implementation to work on Asset List Item View Models.
- Previously, primary image view was not set for standard asset logo display because
the shape was not circle. Now, it's set to the correct shape and size. Also,
PNGImageSource configuration moved to an extension function so that it can be
reused on other types of images.
- Fixed minor layout issues on the list item view.