♻️ Refactor the approach for the alert messages
It had been named as the story sheet before, but alongside the
modifications in 'Macaroon' library, the corresponding changes are
made in the application.

🚸 Fix the sorting dancing for the accounts

The account address is being used now as a fallback choice for sorting
accounts if the preferred choice depends on the USD value of the
accounts. It was need when two different accounts has the same USD value.

🚸 Refine the navigation bar animations on 'Home'

A few minor improvements are made to refine the animations in the
navigation bar on the 'Home' screen.