✨ Implement show more component of asset's about screen
Implemented a view for show more/less texts on the label. Used a label to calculate
line count of the string that fits to the given width. Tried calculating on the view but
the logic was too much to handle on the view. Used a different approach while
calculating the height of the view because the height of the detail label should be
calculated based on the currently displayed text size.