🐛 Fix collectible check in "OptOut" & "TransferBalance"
We were creating `AssetDecoration` from `Asset` object and passing it
to the `OptOutAAssetViewModel` & `TransferAssetBalanceViewModel`.
In both view models we are checking if the asset is collectible
or not and bind the properties according to. But `isCollectible`
check was always failing because `AssetDecoration`'s init from Asset
is setting `collectible` to nil. Fixed it with passing `Asset` directly.