♻️  Adapt to the changes in 'Macaroon' - The background effect of the footer in `ScrollScreen` - The shadows
🚸 Adjust the additional safe area in 'Account Assets List'
If it isn't possible to show the action view to show the quick actions
menu, then the safe area doesn't need to be adjusted for the button
in the 'Account Assets List' screen.

🚸 Improve dequeueing `OptInAssetCell`
The data is just reset when an `OptInAssetCell` is prepared to be
dequeued in the 'Opt-In Asset' screen.