💄 Update the loading states of `Asset Action Confirmation` Previously, a blocking loading indicator was being used. It is changed to a skeleton one in the `Asset Action Confirmation` screen.
💄 Update the bottom content inset in 'ASA Discovery'
If the screen has enough space not to scroll, then the height of the
opt-in/opt-out quick action view doesn't change the bottom content inset
of the scroll view in the 'ASA Discovery' screen.

♻️  Insert 'https' to all urls if any doesn't have a scheme
Since the missing scheme causes issues, crash for `SFSafariViewController`
or not direct to Safari app, `https` is inserted to all web urls before
opening any.

♻️  Clean up the codebase