🐛 Limit ledger account selection for ledger devices
If a rekeyed account is already in the ledger accounts on the same ledger device,
it should not be added to the list again. It was not the case. If there was an account
in the ledger which is rekeyed to the another ledger account on the device list,
the account was shown twice on the list.