🐛 Fix opt-in request modal trigger for watch account
Issue:
When a user has added a watch account, they can also receive opt-in requests
for that account. In case of a 3rd account trying to send a particular asset
that the watch account has not opted in to, we trigger the opt-in request modal.

Fixed with adding watch account check.