🐛 Fix announcement issues
If push notifications are disabled, we weren't sending the device details
when node is changed. It was causing announcements to not be loaded because
we return early If device ID is nil for corresponding node. Also empty
announcement results were not notifying the delegate.
They are fixed with this commit.