🐛 Fix hanging issue on syncing accounts The cancellation of the operation for refreshing the currency wasn't done properly so it was causing the sync queue not to run the upcoming operations in the queue.
🚸 Prevent consecutive calls to sync accounts
The first time the block cycle tracker is being started, it notifies its
listeners instantly, and immediately after it starts to track the next
block cycle which returns quickly, and notifies its listeners again at
short notice which there is no need. So, right now we are passing this
first notification.