🐛 Change WC transaction signer account existence warning
Previously, we were checking the existing account as sender and receiver for the wallet
connect transaction. Changed this implementation to check whether the required signer
exists in the wallet. Instead of checking this issue while parsing the deep linking, moved this
control to the WCTransactionValidator to show the error properly when needed.