👌 Make requested changes for `CollectibleListLocalDataController`
💄 We previously used the layer's mask to check the status of the
shimmer animation. However, this was incorrect because the system
sometimes removes the animation, for example when navigating to another
screen or when the app goes to the background. As a result, the `startAnimating`
function was not executing because we are checking the `isAnimating`
property, which returned `true` when the animation had been stopped by the
system. We were stopping and starting the animation again to fix this issue, but
this is no longer necessary because the `isAnimating` property will now
correctly reflect the status of the animation with checking layer's animation,
and we can simply use the `startAnimating` function.