🐛 Fix collectible list layout issue
There was a problem with the layout of a collectible list in the tab bar
and the Account Detail's NFTs list. The collectible list and
the Account Detail's NFTs list were both trying to use the same layout
objects, which resulted in an unexpected blank screen in the collectible list.
It is fixed.