🐛 Add fallback amount from the quote for the paid and received amounts
We are parsing the transactions and getting the exact received amounts for the swap transactions. It was
suggested to do this way since the amounts slightly can differ from the notifications and the exact transaction
amount on the blockchain. Added the amount on the quote as fallback when there is no account that can be
categorized as received for paid.