🐛 Avoid truncation of URLs in `CollectibleDescriptionView`
When we tap the truncated URL in `CollectibleDescriptionView` component, It
opens the truncated URL instead of actual URL. This issue is normal because
we truncate the text with string prefix method instead of changing the
`lineBreakMode` of the label like `ShowMoreView` component.  It is fixed
by avoiding the truncation of the URLs in the component.