✨ Add loading state for the account detail screen
We needed to add a loading state on the account detail
since there are some operations that needs to be handled
such as filtering, searching and sorting. Added two content update functions, one for the loading with the asset list and
one for the only asset list. Used the first one on the initial launch, search, filter and sort actions. Used the second one
for the block updates.