🐛 Fix grouping transactions to display app calls
We're not able to display app call transactions on the asset
details even if we receive it from the indexer since we filter
the transactions after we receive it. Removed these filters
to depend on directly to the indexer. Kept the all transaction
filters to keep only supported transaction types.