🐛 Fix error on the Swap Flow while opting-in to an ASA
There were a race condition while observing the events of the
coordinator. When user tries to add an asset, we were observing
the event both on the account selection and the asset selection
screens due to the Discover flow. Removed the observers of the
account selection when it's not needed anymore to avoid this
conflict.