♻️ Change multiple transaction request structure
We were not able to display the multiple transaction requests
at the same time properly. Refactored the structure so that we
can display the transaction requests if it's not already displaying
a transaction request from the same topic. Added a rejection
case to handle the session not found case as well.