✨ Handle operations of standard rekeyed accounts properly
Accounts that are rekeyed to another standard accounts was
not handled on the app since it was not possible via any UI and
the some operations like transactions were giving an error even
though the both rekeyed account and authorized account exist
on the device. Handled this issue by using the signer account
while signing the transactions and differentiating rekeyed
accounts from the ledger flows.