♻️ Change transaction signature error controls
We had a control system that checks whether an account can
sign a transaction. It was tied with the BaseViewController so
that it was causing issues. Refactored that protocol so that it
can check whether there is an error related to the signature
of that account and inform the user accordingly by using the
transaction controller. Also, there were two missing parts that
we did not check this error. Updated these places.