🐛 Fix Account Detail NFTs Tab Pending Item Issue
When a pending item is added to the NFT account detail screen
for a specific account while opting in or opting out from an asset
from that account, it is also displayed on other account's account
detail screens. The pending item should only be displayed on the detail
screen of the relevant account. It is fixed.