🐛 Fix "CollectibleDetail" media becomes invisible on status updates
When the collectible status changes due to opt-out/opt-in actions,
the screen is reloaded, which causes the media to become invisible.
It is assumed that after reloading, a new cell is created, and the
old cell's view is removed. Subsequent usages (dequeue) of the old
cell do not show any views. Because we have a check for the
contextView's oldValue, and if the new value equals the old value,
we have a return statement that prevents the views from being added back.