🐛 Finalize the fix for video NFT double play issue
The issue of video (with sound) NFT double play in the 3D screen and
NFT Detail screen, leading to a poor user experience, has been resolved.
Additionally, the problem of videos not starting and stopping correctly
with scroll has also been fixed. Previously, videos would start even if the
cell was not fully centered, as only a portion of the cell was displayed.