✨ Finalize displaying deleted assets
The user was unable to opt out from the deleted asset, which is
impacting his/her minimum balance calculation since we weren't displaying
the deleted assets in the app. With this commit, we will display deleted assets
in the app with a "(Deleted)" notation so that the user will be able to opt out
from them. To determine whether an asset is deleted or not, the "is_deleted"
field from the Pera API should be used, rather than the "deleted" field from the
indexer API.