🐛 Add temporary fix for issue not presenting "WC Transaction" screen
A delay has been added to ensure that the screen is not presented
before the top view controller's view is added to the window's hierarchy.
If the top view controller's view is not in the hierarchy, UIKit will not
present the WC Transaction screen.
Error: Attempt to present <*> on <*> (from <*>) whose view is
not in the window hierarchy.