🐛 Fix rekeyed account's missing ledger detail issue
When we have an account in the wallet that is in a "No auth" state
because the account is imported via a passphrase, but it has an auth address
and its auth account is a Ledger account, importing this account via a passphrase
will not add its rekey detail. However, when we import the Ledger account
(i.e., the auth account), the account state does not change to "Standard to Ledger rekeyed".
from "No auth". This is because we don't add the rekey details to the source
account (rekeyed account), and its state remains fixed. It is fixed.