✨ Remove Device ID from ASB exports
Reason: This can be classified as a security leak. The Device ID is not
being used in our logic at the moment and it would be better to keep it
off the export since it is an optional field. The risk is minor, but it
would be better on the safe side and keep this information private.